src/main/java/com/java_43e/logisticsproject/controller/EmployeeController.java
package com.java_43e.logisticsproject.controller;

import com.java_43e.logisticsproject.entity.Employee;
import com.java_43e.logisticsproject.service.database.EmployeeDatabaseService;
import com.java_43e.logisticsproject.service.impl.ResourceNotFoundException;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/employees")
@RequiredArgsConstructor
public class EmployeeController {

    private final EmployeeDatabaseService employeeDatabaseService;

    @GetMapping("/getAll")
    public List<Employee> getAllEmployees() {
        return employeeDatabaseService.getAllEmployees();
    }

    @GetMapping("/{id}")
    public ResponseEntity<Employee> getEmployeeById(@PathVariable Integer id) {
        try {
            Employee employee = employeeDatabaseService.findById(id);
            return ResponseEntity.ok(employee);
        } catch (ResourceNotFoundException ex) {
            return ResponseEntity.notFound().build();
        }
    }

    @PostMapping("/add")
    public ResponseEntity<Employee> addEmployee(@RequestBody Employee employee) {
        employeeDatabaseService.saveOrUpdateEmployee(employee);
        return ResponseEntity.ok(employee);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Employee> updateEmployee(@PathVariable Integer id, @RequestBody Employee employee) {
        try {
            employeeDatabaseService.findById(id);
            employeeDatabaseService.saveOrUpdateEmployee(employee);
            return ResponseEntity.ok(employee);
        } catch (ResourceNotFoundException ex) {
            return ResponseEntity.notFound().build();
        }
    }
}
